LayoutTests/accessibility/duplicate-hidden-label-text-expected.txt

This test ensures we compute the right string value for unrendered text (with this test exercising one such possible way: hidden labels).

	AXTitle: Foo!
 Bar Baz,
	AXDescription:
	AXHelp:

	AXTitle: Foo!
 Baz,
	AXDescription:
	AXHelp:

PASS successfullyParsed is true

TEST COMPLETE
